+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/ExamplesProperty.scala
@@ -2,7 +2,7 @@ package edu.rit.cs.mmior.jsonoid.discovery
 package schemas
 
 import scala.collection.mutable.ListBuffer
-import scala.math.{exp, floor, log, random}
+import scala.math.{exp, floor, log}
 import scala.util.Random
 
 import Helpers._
@@ -60,10 +60,10 @@ final case class ExamplesProperty[T](
       sampleValue :: examples
     } else if ((totalExamples + 1) <= nextSample) {
       // Use Algorithm L to determine the next sample to take
-      newNextSample += floor(log(random) / log(1 - sampleW)).toInt + 1
-      newSampleW = exp(log(random) / p.maxExamples)
+      newNextSample += floor(log(Random.nextFloat) / log(1 - sampleW)).toInt + 1
+      newSampleW = exp(log(Random.nextFloat) / p.maxExamples)
 
-      val replaceIndex = floor(random * p.maxExamples).toInt
+      val replaceIndex = floor(Random.nextFloat * p.maxExamples).toInt
       examples.slice(0, replaceIndex) ++ List(sampleValue) ++ examples.drop(
         replaceIndex + 1
       )
@@ -108,7 +108,7 @@ final case class ExamplesProperty[T](
       newExamples.length < p.maxExamples && (aIndexes.nonEmpty || bIndexes.nonEmpty)
     ) {
       if (
-        aIndexes.nonEmpty && (bIndexes.length === 0 || random <= sampleRatio)
+        aIndexes.nonEmpty && (bIndexes.length === 0 || Random.nextFloat <= sampleRatio)
       ) {
         newExamples.append(examples(aIndexes.remove(0) - 1))
       } else {